# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "AI Task Team Manager API is a backend service designed to manage team tasks using artificial intelligence.",
        "title": "AI Task Team Manager API",
        "contact": {},
        "version": "1.0"
    },
    "host": "192.168.100.32:8080",
    "basePath": "/",
    "paths": {
...
        "/boards/{boardId}/invite": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Board owner invites a user to the board",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Invite"
                ],
                "summary": "Send invitation",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Board ID",
                        "name": "boardId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Invite request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/invite.CreateInviteRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "invitation sent",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "only the board owner can send invitations",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "already a member / invitation already pending",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/comments/{commentId}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a comment; only the author can delete",
                "tags": [
                    "Comment"
                ],
                "summary": "Delete comment",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Comment ID",
                        "name": "commentId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "deleted",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "access denied",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "comment not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/invites": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns all pending invitations for the authenticated user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Invite"
                ],
                "summary": "List my invitations",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/invite.InviteResponse"
                            }
                        }
                    }
                }
            }
        },
        "/invites/{inviteId}/accept": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Accept a pending board invitation",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Invite"
                ],
                "summary": "Accept invitation",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Invite ID",
                        "name": "inviteId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "accepted",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "access denied",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "invitation not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/invites/{inviteId}/reject": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Reject a pending board invitation",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Invite"
                ],
                "summary": "Reject invitation",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Invite ID",
                        "name": "inviteId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "rejected",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "access denied",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "invitation not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/tasks/{taskId}/comments": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns all comments for a task; caller must be a board member",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Comment"
                ],
                "summary": "Get task comments",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Task ID",
                        "name": "taskId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/comment.CommentResponse"
                            }
                        }
                    },
                    "403": {
                        "description": "access denied",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "task not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Add a comment to a task; caller must be a board member",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Comment"
                ],
                "summary": "Add comment",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Task ID",
                        "name": "taskId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Comment body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/comment.CreateCommentRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/comment.CommentResponse"
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "access denied",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "task not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/upload/photo": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Upload a profile photo (max 5 MB, jpeg/png/gif/webp). Returns a URL to use in employee fields.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Upload"
                ],
                "summary": "Upload photo",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Photo file",
                        "name": "photo",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "url",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "comment.CommentResponse": {
            "type": "object",
            "properties": {
                "author_id": {
                    "type": "integer"
                },
                "content": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "task_id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "comment.CreateCommentRequest": {
            "type": "object",
            "required": [
                "content"
            ],
            "properties": {
                "content": {
                    "type": "string",
                    "minLength": 1
                }
            }
        },
...
        "invite.CreateInviteRequest": {
            "type": "object",
            "required": [
                "user_id"
            ],
            "properties": {
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "invite.InviteResponse": {
            "type": "object",
            "properties": {
                "board_id": {
                    "type": "integer"
                },
                "board_name": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "invitee_id": {
                    "type": "integer"
                },
                "inviter_id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter the token with the `Bearer ` prefix, e.g. \"Bearer abcde12345\".",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}